Expand OperatorAssumptions to allow for choosing conditioning
This is the last thing of https://github.com/SciML/LinearSolve.jl/issues/283 to be fixed up. It allows for different levels of conditioning assumptions, in order to allow for making a trade-off between performance and numerical stability. Right now I've handled the very ill-conditioned cases but not the well-conditioned case (which is the case the issue is interested in).