Mutable cache form
I think this is required to fix https://github.com/SciML/LinearSolve.jl/issues/302 . Otherwise there is no way to set some kind of "has been factorized". The reason for the functional immutable form was to be easier on the compiler, but I don't think that has worked out in the end. It has been a bit unergonomic to have to keep finding the latest cache, and in scenarios where it's kept longer (i.e. `init` the ODE solver) it will allocate just the same, while when it's in a more constrained context (`solve` in the ODE solver) the compiler will elide the allocation for both the mutable and immutable forms these days. Also, this puts LinearSolve more in line with the rest of the SciML interface libraries.

Fixes https://github.com/SciML/LinearSolve.jl/issues/181and fixes https://github.com/SciML/LinearSolve.jl/issues/199