WIP: Make more libraries into an extension?
I think we should be considering some kind of split to LinearSolveCore vs LinearSolve, a la OrdinaryDiffEq. The default algorithm requires RecursiveFactorization and MKL, but we could definitely make a version that doesn't have these deps. But at least sparse stuff can go into an extension since those always dispatch on the existance of sparse matrices.

This PR isn't quite complete yet, more needs to be removed with sparse, but it shows the right direction.