Safety measures around AppleAccelerate (#676)
```julia
Enzyme Derivative Rules: Error During Test at /home/runner/.julia/packages/SafeTestsets/raUNr/src/SafeTestsets.jl:30
  Got exception outside of a @test
  LoadError: could not load library "/System/Library/Frameworks/Accelerate.framework/Accelerate"
  /System/Library/Frameworks/Accelerate.framework/Accelerate.so: cannot open shared object file: No such file or directory
  Stacktrace:
    [1] fix_ptr_lookup(name::String)
      @ Enzyme.Compiler.JIT ~/.julia/packages/Enzyme/LMVya/src/compiler/orcv2.jl:63
    [2] add!(mod::LLVM.Module)
      @ Enzyme.Compiler.JIT ~/.julia/packages/Enzyme/LMVya/src/compiler/orcv2.jl:251
    [3] _link(job::GPUCompiler.CompilerJob{<:Enzyme.Compiler.EnzymeTarget}, mod::LLVM.Module, edges::Vector{Any}, adjoint_name::String, primal_name::Union{Nothing, String}, TapeType::Any, prepost::String)
      @ Enzyme.Compiler ~/.julia/packages/Enzyme/LMVya/src/compiler.jl:5661
    [4] cached_compilation
      @ ~/.julia/packages/Enzyme/LMVya/src/compiler.jl:5750 
```

Basically, you can't do option dependencies, but BinaryBuilder just simply sends no binary if the platform isn't supported. So LinearSolve.jl always has a dependency on AppleAccelerate, and if the binary exists it defaults to it (since it's pretty much always the fastest on M-series mac) and otherwise it disables it. Enzyme still seems to want to try to call it, so this should fix that.