prepare to properly suppory lazy libraries (#789)
The is_available checks should be runtime checks, which will be possible
to handle fully properly once MKL_jll and OpenBLAS_jll are upgraded to
support that. For now, just convert the variables here so that they
aren't trying to do a runtime lookup of a module that was not imported.