Better safeguard on unloadable MKL (#793)
* Better safeguard on unloadable MKL

Tries to safeguard https://github.com/julia-actions/setup-julia/pull/300 https://github.com/JuliaLang/julia/issues/55878 where the user has requested an incorrect architecture and thus we may believe we can load MKL until we check `MKL_jll.is_available() == false`. The reason is because we're being lied to about `@static if Sys.ARCH === :x86_64 || Sys.ARCH === :i686` which then makes it take this branch and then MKL does not exist. We could instead check `isapple` another type of workaround here and 99% of users would be perfectly fine, but if this is safe this is slightly better for the 1 remaining Intel Apple user that exists in the wild somewhere.

* Update LinearSolve.jl